LPS-61082 Make ConvertDatabase extensible through OSGi services that override the interface ConvertDatabaseProcess
